Instance sync decomposed into services
