+++ src/main/kotlin/com/cognifide/gradle/aem/instance/InstanceService.kt
@@ -1,0 +1,10 @@
+package com.cognifide.gradle.aem.instance
+
+open class InstanceService(val sync: InstanceSync) {
+
+    val aem = sync.aem
+
+    val instance = sync.instance
+
+    val project = sync.project
+}
