+++ src/main/kotlin/com/cognifide/gradle/aem/pkg/tasks/PackageDelete.kt
@@ -23,8 +23,8 @@ open class PackageDelete : PackageTask() {
         aem.progress(instances.size * packages.size) {
             aem.syncPackages(instances, packages) { file ->
                 increment("${file.name} -> ${instance.name}") {
-                    val pkg = getPackage(file)
-                    deletePackage(pkg.path)
+                    val pkg = packageManager.getPackage(file)
+                    packageManager.deletePackage(pkg.path)
                 }
             }
         }
