+++ src/main/kotlin/com/cognifide/gradle/aem/pkg/tasks/PackageInstall.kt
@@ -16,8 +16,8 @@ open class PackageInstall : PackageTask() {
         aem.progress(instances.size * packages.size) {
             aem.syncPackages(instances, packages) { file ->
                 increment("${file.name} -> ${instance.name}") {
-                    val pkg = getPackage(file)
-                    installPackage(pkg.path)
+                    val pkg = packageManager.getPackage(file)
+                    packageManager.installPackage(pkg.path)
                 }
             }
         }
